Changes to the way brush toggles work
There was a "bug" where the default bush was named `Shroud` but when
toggled the name would change to `Shadow`. The use of `Shadow` has been
fixed.

Furthermore, this change also corrects brush behaviour. Previously the
button behaviour was to:

* Show `Shroud Brush` when brush mode clears FOW
* Show `Clear Brush` when brush mode adds FOW
* Show `Square Brush` when DM uses a circular brush
* Show `Circle Brush` when DM uses a square brush

Since this seemed counter intuative, the toolbar shows the brush
currently in use:

* Show `Shroud Brush` when brush mode adds FOW
* Show `Clear Brush` when brush mode clear FOW
* Show `Square Brush` when DM uses a square brush
* Show `Circle Brush` when DM uses a circular brush